vpp: fix video afbc display for 8g [1/1]
PD#SWPL-45779

Problem:
fix video afbc display for 8g

Solution:
fix video afbc display for 8g

Verify:
t7

+++ include/linux/amlogic/media/vfm/vframe.h
@@ -437,8 +437,8 @@ struct vframe_s {
 
 	u32 canvas0Addr;
 	u32 canvas1Addr;
-	u32 compHeadAddr;
-	u32 compBodyAddr;
+	ulong compHeadAddr;
+	ulong compBodyAddr;
 
 	u32 plane_num;
 	struct canvas_config_s canvas0_config[3];
